VideoDao renamed to VideoRepository
