+++ backend/src/main/java/lt/vu/tube/repository/VideoRepository.java
@@ -1,4 +1,4 @@
-package lt.vu.tube.dao;
+package lt.vu.tube.repository;
 
 import lt.vu.tube.entity.Video;
 import org.springframework.data.repository.CrudRepository;
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.UUID;
 
 @Repository
-public interface VideoDao extends CrudRepository<Video, UUID> {
+public interface VideoRepository extends CrudRepository<Video, UUID> {
 
 }
